requirements.txt
streamlit==1.25.0
python-dotenv==1.0.0
google-generativeai==0.2.0
youtube-transcript-api==0.5.0
googletrans==4.0.0-rc1
